buildsys: use automake
This introduces a `Makefile.am`, replacing `Makefile.in`. We get a few
benefits from this:

- `dist` and `distcheck` targets.
- Portability, especially to non-GNU systems.
- Simplification.

Automake's policy is that of explicitness; thus, the `Makefile.am` is
long due to the `EXTRA_DIST` included for the tests. This comes into
play in the `distcheck` target, which assumes the tests are distributed
with the program so that package builders can ensure that the system
works locally.

The `check` target makes use of the existing test suite and
infrastructure; little has changed from `Makefile.in`.

The source.mak file existed, but various variable names within conflicted
with automake. Rename those and remove the remaining unused ones.

While here, remove generated files from our commits.

Set foreign in AM_INIT_AUTOMAKE in configure.ac as suggested by @b4n.