Generate Makefile from autogen.sh
The `EXTRA_DIST` is the files needed for the tests. From autogen.sh,
generate a Makefile which is included into testing.mak, listing all the
files required for the tests.

Instead of deprecating autogen.sh, we bring it back by updating the
documentation and build scripts to make use of it.

While here, fix some more automake complaints from source.mak.