Clean up old-docs/INSTALL
This is the old installation instructions. Atop it has a summary of the
important configure arguments, and below it has the standard GNU
installation instructions.

There is no quick and simple way to rename these programs, but there is
a simple way to add a suffix to the end. Document that and remove the
other documentation.