main: introduce sInpufFileInfo to handle `input' and `source' in the same way
This commit is a part of --line-directive overhauling.

line-directive overhauling policy
---------------------------------

When #line directive is found in an input file, the low-level
input handling of exuberant-ctags treats a file(a source
file) specified in the directive as the input file when
--live-directive option given. As the result the source file
name is used instead of the input file name in tags file.
The information of the input file name is overwritten with
the source file name. exuberant-ctags generates tags either
for the input file name or  the source file name.

universal-ctags will do different; it will record the raw/low
information. Upper layer tools can do something interesting
with the information. universal-ctags uses the input file
name anytime.  If --line-directive is given, universal-ctags
generates tags both for the input file name and source file
name.

About this commit
---------------------------------

This commit introduces sInpufFileInfo structure which was
defined inside sInputFile as sSource for recording source
file information. sInpufFileInfo will be used for keeping
information both input and source files.

In this commit only name and lineNumber fields of input.File.input
are used.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>