Change the type for tagPath field of inputFileInfo to vString
When --line-directive=yes is specified, inputFile.source.tagPath may
be stored to a tagEntryInfo instance as a pointer.

Consider #line directives appear more than once like:

    #line 20 foo.input
    #line 30 bar.input

"foo.input" pointed by inputFile.source.tagPath may be stored to a tagEntryInfo.

When ctags finds the line of "bar.input", in current implementation,
ctags destroys a buffer "foo.input" via inputFile.source.tagPath is freed, then
ctags allocate and fill new buffer for "bar.input".

In other hand there are possibilities that a tag entry keeps an address pointing the "foo.input"
buffer. The buffer is destroyed without notifying it to the tag entry when ctags read the
line of "bar.input".

Solve this issue in three steps.

(1) converting the type of tagPath field to vString (this commit),
(2) keeping the buffer to a stringList instead of destroying the buffer that tagPath points, and
(3) destory the stringList when all tags are processed for an input file

The step (1) is needed to store a buffer pointed by tagPath field to the stringList.

Commits for step (2) and (3) follow this commit.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>