main: remove unused parameter from renderRole function
To avoid circular header file inclusion, tagEntryInfo was defined in
multiple headers. The one of definition is needed in a parameter of
prototype declaration of renderRole function. However, the parameter
is not used in the function now.

This commit removes both the parameter and the one of tagEntryInfo type
info definition.

This is needed to build ctags on RHEL5.11.
Close #699.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>