buildsys: embed the current git commit id(or tag) to ctags binary
If configure is used for building and building is done in
a git repository, git commit id is embedded to ctags
executable (as before). The id can be seen in the output
of `ctags --version' or in __TAG_PROGRAM_VERSION pseudo tag
in a tags file.

If ctags built on a platform where configure is not used,
the id is set to 0. If ctags at the outside of a git repository,
the id is set to 0.

This feature will be used in regression test I'm designing now.

For tagged version, embed the tag instead of commit id.
Suggested by @k-takata.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>