main: Introduce an option to add "UCTAGS" as prefix to field names
We want to add more fields to tags output.

e.g. "returnType:" of a function declaration,
"decorator:" or "annotation:" attached to a name.

In other hand there are many tags generators. .e.g.  coffeetags and
ripper-tags. The authors of these tools may want to add the
generator(and its target language) own new fields to the output.

There is concern about confliction of field names.

For providing a workaround of the confliction, this
commit introduce --put-field-prefix option.

When --put-field-prefix is given, "UCTAGS" is prefixed to the name of
all fields newly introduced in universal ctags.

Example output(--list-fields):

    [yamato@x201]~/var/ctags-github% ./ctags --put-field-prefix --list-fields | grep UC
    r	UCTAGSrole	role	format-char	off
    Z	UCTAGSscope	Include the "scope:" key in scope field(use s)	NONE	off
    [yamato@x201]~/var/ctags-github% ./ctags --list-fields | grep '^[rZ]'
    r	role	role	format-char	off
    Z	scope	Include the "scope:" key in scope field(use s)	NONE	off

Example output(--extra=+r --fields=+r):

    [yamato@x201]~/var/ctags-github% cat /tmp/foo.h
    #include <stdio.h>
    [yamato@x201]~/var/ctags-github% ./ctags -o - --extra=+r --fields=+r /tmp/foo.h
    stdio.h	/tmp/foo.h	/^#include <stdio.h>/;"	h	role:system
    [yamato@x201]~/var/ctags-github% ./ctags --put-field-prefix -o - --extra=+r --fields=+r /tmp/foo.h
    stdio.h	/tmp/foo.h	/^#include <stdio.h>/;"	h	UCTAGSrole:system

Signed-off-by: Masatake YAMATO <yamato@redhat.com>