buildsys: make an empty repoinfo.h when building ctags from tarball
repoinfo.h was general even when building ctags from tarball.  As
reported by @hauleth in #727, code snippet in Makefile.am made broken
repoinfo.h. This commit improves the situation.

The code snippet in Makefile.am in this commit make an empty repoinfo.h.
Instead CTAGS_REPOINFO is defined as NULL in repoinfo.c if CTAGS_REPOINFO
is not defined.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>