Make: rename "include" kind and "generic" role
"include" was used as a kind name for a tag on included makefile.
Now I recognized "include" is not good kind name.
"include" is an action, it just implies what the associate tag is.
But a kind name, "makefile" is more explicit.

For a reference tag for include makefile, "generic" is used as a role
name. It just implies the role; a user can know a makefile referred in
a makefile will be included. But a role name, "included" is more explicit.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>