Sh: rename "source" kind and "generic" role
"source" was used as a kind name for a tag on load scripts.
Now I recognized "source" is not good kind name.
"source" is an action, it just implies what the associate tag is.
But a kind name, "script" is more explicit.

For a reference tag for loaded scripts, "generic" is used as a role
name. It just implies the role; a user can know a script referred in a
script will be included. But a role name, "loaded" is more explicit.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>