main, refactor: pass parserDefinition instance to its parser callback function
In the original ctags architecture, parser callback function of
parserDefinition, the parserDefinition instance, and all the state of
parsing are separated.

This commit is initial trail for integrating them in object oriented way;
parserDefinition is considered as an object, which has the most
of all the sate of parsing as members of the object. The parser callback
is considered as a method of the object. Inheritance and Mix-In can be
introduced in the future.

My analysis of the original architecture:

    parser callback function of parserDefinition took no argument.  If
    more than two parserDefinitions share the same callback function,
    isInputLanguage macro must be used to know the owner
    parserDefinitions. findCTags function defined in parsers/c.c is
    one of such functions. findCTags is shared in CParser, DParser,
    CppParser, CsharpParser, JavaParser, and VeraParser.  In addition
    all the state of parsing are stored to file local variables.

In a parserDefinitions there were two fields specifying the parser
callbacks: one for one path parsing, and the other for multiple path
parsing. This commit integrates the two, too.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>