Python: improve handling multiline keyword arguments
Close #750.

@strokirk wrote:

Multiline keyword arguments in python are parsed as variables by
ctags. They should not be included in the generated tags file.

For example, this input:

    class Foo(object):
	def bar(self,
		default=None):
	    pass
	baz = dict(
	    argument=None,
	)

    variable = dict(
	keyword=None,
    )

Results in this output:

    Foo test.py /^class Foo(object):$/;"    c
    argument    test.py /^        argument=None,$/;"    v   class:Foo
    bar test.py /^    def bar(self,$/;" m   class:Foo
    baz test.py /^    baz = dict($/;"   v   class:Foo
    keyword test.py /^    keyword=None,$/;" v
    variable    test.py /^variable = dict($/;"  v

Unexpectedly `argument' and `keyword' are captured.

This commit adds code that skips next pairs of parenthesis after finding
a variable.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>