buildsys: use pattern rule instead of suffix rule
To run optlib2c translator, a suffix rule(.c.ctags) was used in
Makefile.am. However, suffix rules cannot handle extra dependency
like:

       .ctags.c: $(OPTLIB2C)

The extra dependency runs the translator when the translator itself is
update.

This commit replaces the suffix rule with a pattern rule.
With the suffix rule, the extra dependency can be added:

     %.c: %.ctags $(OPTLIB2C)

Signed-off-by: Masatake YAMATO <yamato@redhat.com>