lxpath: add XML generic selector
Many of XML meta parser based parsers may have the same extension: xml.
New selector, selectByDTD is for selecting a proper parser for the xml file.

This implementation is not efficient; it crates an dom tree for selecting.
The dom tree should be reused in the selected parser.

This implementation is ugly; per concrete parser rules for selecting are
hard-coded in selectByDTD. These rules should be in parserDefinitions.

Making a XML dom tree from input memory area should be generalized.
So it can be used another areas of ctags.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>