units: introduce a marker for skipping a test case if ctags has a specified feature
Units/TEST/features was introduced to specify features needed to run
the TEST case.

This commit introduces a marker for skipping a test case if ctags HAS a
specified feature; the test is skipped if a line of the file is
started with ``!`` marker, and a target ctags have the feature
specified with ``!``.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>