Python: fix a bug about multi-line keyword arguments
Close #765.

A bug description quoted from #765 reported by @m-novikov:

    Function calls with multi-line keyword arguments results in
    keyword arguments parsed as variables. Given:

    foo(
	    arg1=1,
	    arg2=2
    )
    def bar():
	    pass

In the example arg1 and arg2 are captured as tags.

This commit fixes this bug.

Python parser processes input file line by line.  Line may include a
tag or not. Python parser applies rules for capturing a tag to the
each input line. This commit introduces a fallback rule which calls
skipParens. As the result foo (...) in the above example are skipped.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>