C++: handle lesser and greater operators in template parameters
Close #769.

`<' and `>' operators can be used in parameters list of template
like:

    template<char i, bool B = (i>0)>     void f0(void) { }
    template<char i, bool B = i < 10>    void f1(void) { }

ctags could not handle them well.

This commit tries to improve the way to handle them.

The original checks only the balance of angle brackets.
In addition the original way, this commit considers
following things when checking the balance of angle
brackets:

1. lesser and greater operators between rounded brackets are ignored.

2. When `=' is found, lesser and greater operators are
   ignore till the input satisfies one of following conditions:
   2.1 `,' is found outside rounded brackets, or
   2.2 `>' is found outside rounded brackets,

Signed-off-by: Masatake YAMATO <yamato@redhat.com>