main: introduce fallback mechanism in language selection
ctags tries variety ways(HINTS) to choose a proper language for given
file Name. If multiple candidates are chosen in one of the hint, a
SELECTOR common between the candidate languages is called for
selecting one from the candidates. However, the SELECTOR can
selects none. In this case all candidates chosen by the HINTS
were never used; and the input is not parsed.

The basic assumption of this commit is that running one of the
candidates for parsing the input will be better instead of doing
nothing about the input. With this commit candidates are recorded
before invoking a SELECTOR. If the SELECTOR selects none, use the
one of the candidates as a fallback parser.

The change will be tested with --language=- option. The option
can make the SELECTOR selects none. The test case will prepared
in another commit.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>