main: sort the parser candidates by their names
For an input file, ctags nominates more than one parsers as
candidates. Patterns and extensions are used for nominating.
Candidates are stored to an ARRAY. Then ctags selects one of them if
"selector" common between candidates is available.

If a selector is not available or the selector doesn't work as
expected, ctags selects the first one of candidates in the array.
(fallback rule)

A commit

   main, INCOMPATIBLE: give patterns higher priority than extensions
   when choosing a parser

introduced new code that sorts the array.

With sorting, the candidates nominated by patterns get higher priority
than the candidates nominated by extensions.

Issue this commit deals is about this sorting. qsort standard C
function used in the sorting doesn't do stable sorting. This means the
result of the fallback rule, selecting "the first one of candidates in
the array", cannot be predictable.

This commit modifies the comparing
function(sortParserCandidatesBySpecType) of qsort; if two parsers has
the same priority, the function compares their name with strcasecmp.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>