Python: improve the handling of doc quotes used in a list of variable line
Partially close #781.

Code handling input lines including variable
assignments didn't take doc quotes into account.

    1:    var=("""
    2:    str
    3:    """,)
    4:    anothervar=1

The code just parses one of three lines:

    1:    var=("""

The rest are handled in the top level parser code:

    2:    str
    3:    """,)
    4:    anothervar=1

The top level parser code takes doc quotes; it
recognizes the line

    4:    anothervar=1

is in doc quotes. As the result `anothervar' is not
tagged.

This commit fixes this bug partially.

As far as a list is not assigned to a variable in the input line, the
code handling the variable assignments passes the control to
`find_triple_start' in this commit. As the result the python parser
gets the chance to recognize

    1:    var=("""

as the start of doc quotes. When parsing the line:

    3:    """,)

python parser can know the line as the end of doc quotes.
As the result `anothervar' is tagged.

This commit is incomplete. Python parser doesn't handle doc quotes
used as default a parameter:

    def f(var="""
    str
    """):
	    pass
    anothervar=1

Signed-off-by: Masatake YAMATO <yamato@redhat.com>