main, refactor: Remove toStdout special case for opening tag file
It looks like this was a special case to avoid opening the tag file
again. Since the result is going to be written to stdout, the input file
won't be overwritten, so they could skip the open. Let's simplify this,
and just always open the file.

This is part of the effort to make TagFile static to entry.c.