main, php: add a root separator to top level items
makeQualifiedTagEntry() builds a full qualified tag for
a given tag. It uses scope information; put a separator
between upper item and the current item specified with
the given tag.

This commit handles the case when there is no scope information.  If
such case a root separator must be put as prefix to the current item.

e.g.

input:

    <?php
    namespace foo;
    >

the output with --extra=+q:

    \\foo	input.php	/^namespace foo;$/;"	n
    foo	input.php	/^namespace foo;$/;"	n

Signed-off-by: Masatake YAMATO <yamato@redhat.com>