SQL: handle unexpected tokens in declare block
The original code expects a semicolon in a declare block.  However,
as a test case shows, a block can have a token other than a semicolon
at least in pgSQL.

This commit adds an additional rule to stop finding the semicolon for
handling the situation. When 'begin' or 'end' is found, the sql parsr
stops finding the semicolon and exits from declare block handling.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>