Tmain: ignore OldC and OldC++ parser in list-roles.d test case
When introducing newly rewritten parser, we would like to provide the
both new parser and old parser for debugging and providing migration
period to users. In such case the prefix "Old" will be used to the
name of old parser. The old parser should be ignored in this test
case.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>