main: ignore separator fields for KIND_WILDCARD when finding a root separator
KIND_NULL is used as the kind of parent item when finding a root separator of
given kind. separator fields for KIND_WILDCARD was matched with KIND_NULL.
As the result, a kind having KIND_WILDCARD in its separator fields was
recognized as a kind which has a root separator. This should not be.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>