Introduce MIO library that uses file-like operations to access memory
MIO uses two backends - one file-based and one memory-based and a single
API that mirrors the standard C file operations and that can be used
regardless of what backend is used. Using this instead of direct file
operations will allow us to load parsed files into memory first and
then perform file-like operations on memory which is slightly faster.
In addition, this is one of the steps towards making ctags a library
as this will allow the library users to parse contents of their buffers.

In principle the patch is trivial - it just replaces FILE with MIO
pointers fpos_t with MIOpos and uses mio_* file-like operations.