Remove fgetpos() and fsetpos() implementations
I'm not aware of any system where fgetpos() and fsetpos() could be
missing - they are part of the C standard library and should be
available everywhere.

If we wanted to keep them, they should be moved to MIO but as I think
the native implementations are always available, this seems to be
unnecessary so this patch removes them.