Allow typeRef[0] == NULL and generate "unknown" for it in the tags file
At the moment the value of typeRef[0] when no referenced type exists is
parser-specific. The cxx parser uses "typename" in this case, the php
parser uses "unknown". Instead, allow typeRef[0] to be set to NULL and
generate "unknown" for it when writing out the tag entry. This will make
the output identical for all languages.