main: introduce type: field
type: field is introduced in this commit.
The value for the field is passed via typeRef[1] of
tagEntryInfo from a parser to the main part.

To emit the string pointed by typeRef[1] as type:
field, typeRef[0] must be NULL and the field is enabled
with --field=+T. Here 'T' is the letter for type:
field.

The field name and help string are suggested by @techee.
Signed-off-by: Masatake YAMATO <yamato@redhat.com>