main: introduce attachParserFieldToCorkEntry
For attaching a value to a parser own field of an entry on corkqueue,
the memory object for value must be allocated dynamically because the
entry will be free'ed after emitted to tags file.

This means a parser has to call eStrdup before callling attachParserField.
This newly introduced function suppresses the step.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>