main: allow parsers calling makeTagEntry with a tag disabled kind
Old behavior: Assertion failure when a parser calls makeTagEntry with
a tagEntryInfo which associated kind is disabled.

That means a parser must verify "enabled/disabled" of the kind
associated to a tagEntryInfo before calling makeTagEntry.

However, people may forget writing such verification.

With this commit, makeTagEntry just skips the entry silently; it
doesn't emit the passed tagEntryInfo if its kind is disabled.

In some case checking "enabled/disabled" of the kind is still
meaningful for optimization purpose; if a kind is disabled, parser
doesn't have to take time to record the information categorized to the
group.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>