main: move printf function at the end of block when printing scope information
To reuse the code making scope information, I'm working on separating
the code making scope information and the code printing them.
This is the initial step.

A buffer pointed by full_qualified_scope_name was freed the original
code.  However, it is always freed when flusing the cork queue. So
this commit deletes the code freeing the buffer.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>