main: introduce lineNumberOrigin field
For implementing nested input stream, this field is needed.

When a parser makes a tag in a nested input stream, the parser
may want to record the line number of the base input stream,
not the nested input stream. To satisfy the requirement,
the lineNumber of the nested stream must be adjusted in the
stage of initializing the nested stream. The lineNumberOrigin
will be used in this stage.

The value of lineNumberOrigin field is set to lineNumber field
of inputFileInfo when resetInputFile is called.

Before introducing the lineNumberOrigin, the lineNumber field is
set to 0 when resetInputFile is called on the input stream.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>