CXX: separate the data structures related to roles between C and C++
The data structures related to roles were shared between C and C++ parsers.
This can leads inconsistent behavior between role control and kind control
in C and C++.

Currently ctags doesn't provide the way to enable/disable a role, so the
sharing them has no visble impact. However, when ctags provides the way
to do it, follwoing behavior will emerges.

If a user enables a role in a language, a role having the same name in
antoher language is also enabled. The same behavior can be found when
disabling.

An example:

	$ ./ctags --_list-roles
	...
	C	h	system	system header	on
	...
	C++	h	system	system header	on

If a user disables `system' role of C parser, `system' rolf of C++
parser is also disabled.

The behavior itself are not bad. In some case users may want the
above behavior. If a user think c pre-processor is a lower common language
of C and C++, the user may want to enable/disable at the layer.

However, this behavior are inconsistent with how kinds are controlled
between two languages.  Data structures related to kinds are NOW
separated in C and C++. Enabling a kind in a parser has no effects ont
the kind having the same name in another language.

An example:

    $ ./ctags --kinds-C=+f --list-kinds=C  | grep ^f
    f  function definitions
    $ ./ctags --kinds-C=+f --list-kinds=C++  | grep ^f
    f  function definitions
    $ ./ctags --kinds-C=-f --list-kinds=C  | grep ^f
    f  function definitions [off]
    $ ./ctags --kinds-C=-f --list-kinds=C++  | grep ^f
    f  function definitions

Though `f' kind is disabled in C parser, `f' kind of C++ is still enabled.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>