main: verify whether a field is fixed or not in enableField
Basically this is a refactor. However, some warning messages are
changed.

enableField and isFieldFixed exist as separated functions.
This commit merges them.

Before enabling/disabling a field, verification whether
the field is fixed or not in enableField.

enableField doesn't disable any fixed fixed.
enableField makes a warning when a client tries disabling
a fixed field if warnIfFixedField, the 3rd argument is true.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>