output, refactor: simplify an extension field rendering
Introduce renderExtensionFieldMaybe and use it if possible.

renderExtensionFieldMaybe does all (1) checking whether
given extension field is enabled or disabled, (2) checking
whether a value for the field is attach or not, and (3)
printing.

Ideally renderExtensionFieldMaybe can render all the extension fields.
However, for compatibility, we have to maintain old no normalize
output format.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>