python: Fix parsing of multiple variable assignations
When parsing multiple variable assignations, properly recognize the
type of the assignment (value or lambda).

Also, don't erroneously output tags for mere multiple variable
references (i.e. a line consisting solely of `a, b, c`), although it's
highly unlikely to appear in real world code as it would effectively be
a no-op.