main: don't include `^' in a pattern if a tag is at the start of a nested input stream
A nested input stream can start with byte offsets from a line
of the base input stream. In such case a pattern for a tag taken
from the nested input stream should not start with `^'.

Newly introduced getNestedInputBoundaryInfo() tells
makePatternStringCommon() whether it can prepend `^' to a pattern
for the current tag or not.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>