main, refactor: rename get.[ch] to lcpp.[ch]
The original file names get.[ch] are too generic.
I wonder who can image the files handle C processor directives from the file name.

I change it to lcpp.[ch]. The prefix `l' means the file is for
routines useful for implementing a parser for specific area of
`l'anguages. I was afraid that the name conflicts with C++. However,
we found a good prefix cxx for C++.

This naming scheme follows lregex.c, a file originally available in
exuberant ctags source tree. (The interpretation of `l' is my own, not
written explicitly in documents of exuberant ctags.)

Signed-off-by: Masatake YAMATO <yamato@redhat.com>