main: disable "scopeKind" field by default
The field was enabled by default. I wonder why I did so.

This has no impact both tags output and xref output.
In tags output, the field is never used in its printing routine.
In xref output, the field is never printed if the field is not
specified in --_xformat option.

However, this has an impact on planed json output. The field was
printed though it was not enabled in the command line because it was
enabled by default.