jscript: Reject newlines inside a regex literal
There cannot be a newline in a regex literal, so rejecting it improves
handling of malformed or incomplete code.

It is particularly interesting to avoid this because a regular
expression starts the same as a single-line comment, so can easily
appear in incomplete files (e.g. in files currently being edited).

See 11.8.5 "Regular Expression Literals" in
http://www.ecma-international.org/publications/files/ECMA-ST/Ecma-262.pdf