lcpp, OldC, Cxx, DTS, Vera: enable header kind by default
"header" kinds for OldC, Cxx, DTS, Vera were disabled by default.
However, a tag for the kinds is printed always because lcpp calls
makeTagsEnty without verifying the value of `enabled' field of the
kind.

This commit does two things:

(1) verifying the value of the `enabled' field of the "header" kind in lcpp, and
(2) enabling the "header" kinds by default.