main: new dependency type, subparser
The subparser dependency can be used for initializing
a parser when another parser is initialized.

Following example is taken from autoconf parser
under development.

    extern parserDefinition* AutoconfParser (void)
    {
	    parserDefinition* const def = parserNew("Autoconf");
	    ...
	    static parserDependency dependencies [] = {
		    { DEPTYPE_SUBPARSER, "M4" },
	    };

	    def->dependencies = dependencies;
	    ...
	    def->initialize = initializeAutoconfParser;

This code asks the main part of ctags to initialize Autoconf
parser when M4 parser is initialized.

As the name shown, this dependency is useful if you want to register a
parser as a subparser to a host parser in the initializer.

    static void initializeAutoconfParser (const langType language)
    {
	    AutoconfM4Client.lang = language;
	    registerM4ParserClient ("M4", &AutoconfM4Client);
    }

Signed-off-by: Masatake YAMATO <yamato@redhat.com>