main: clear the language stack when closing or resetting input
This commit allows unbalanced pushLanguage.

Basically a language type is pushed, it must be popped.  However,
there is a case that just pushing only operation is preferable; "push
the language" can be used for specifying the name of language printed
in a file tag enabled with --extra=+f.

An example will be found in Autoconf parser which is under developming.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>