main: Implement iFileGetLine() using mio_gets()
This avoids many calls of mio_getc() and in addition eliminates many
calls of mio_getsvStringPut() which get expensive when called for every
single character.

When tested on the new HTML parser, this change increases the speed to
87MB/s from 70MB/s on my machine. When modified the parser so it just
calls getcFromInputFile() until EOF, the speed gets up to about 166MB/s
from about 110MB/s. When profiled, iFileGetLine() almost disappears from
the profile and the CPU usage is distributed between getcFromInputFile()
and mio_gets() (because these two have to read every character). Other
parser's speed improvement depends on how much they are optimized -
the faster the parser is, the more it will benefit from this change.

The greatest part of the patch is the change of the code dealing with
line directives to read characters from a string instead of MIO input
but this change is mostly straightforward.

This patch introduces a small change in how \0 characters are treated
in the input file; before, \0 characters were silently skipped, after
this patch the rest of the line after \0 is skipped. This is unavoidable
because gets() doesn't return the number of characters read. As a
result c-heapoverflow-sh-bug-363 test result has been changed to follow
the new \0 handling.