Rename analyzeToken() and optimize it
The naming of the function is poor - it's basically the same as
lookupKeyword(), it just lowercases the string first. This function
is used for case-insensitive languages.

Introduce lookusCaseKeyword() which is used instead of this
function and which makes an case-insensitive match. This requires
that the hash function is modified so it always lowercases the
input and case-sensitive or -insensitive matching is performed
using either strcmp or strcasecmp() in lookupKeywordFull().

This also avoids string allocation; this is a perfromance killer
especially for parsers which lookup a keyword for every found identifier.
For instance this doubles the speed of the new Python parser (from
17MB/s to 34MB/s on my machine).