Make case-sensitive languages use lookupKeyword() instead of lookupCaseKeyword()
I assume that and Object instance in javascript was supposed to be a
class tag (at least it seems to be the intention of the code). This
means that the corresponding test case had to be modified (because
of lowercasing keywords, Object was never found as a keyword in the
previous code).