units: trap INT signal in subshell
Close #1100.

It seems that a subshell launched in units test
harness exits with zero even if it receives
INT signal. As the result though a user presses
^C from the terminal where "make units VG=1" runs,
the execution of test cases continues. (The execution of
the current test case is stopped but the next one is
started by the test harness.)

This change makes the subshell exits with non-zero when
the subshell receives INT signal.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>