Add generic pointer array and implement strList using it
A resizable array that can be also used as a stack is a useful
data structure that is currently missing in ctags and it's
re-implemented at various places unnecessarily (searching for
'realloc' in ctags sources reveals several places where
pointer array could be used).

The implementation just moves all the non-string-specific
functions from strList into the new ptrArray and uses ptrArray
internally. The only functional changes are:

- instead of creating 0-sized array at the beginning, the
  new implementation creates an array of size 8
- the new implementation always doubles the size of the array
  on reallocation; previously the array was increased by 10
  elements only
- the simplest functions were rewritten as macros to avoid
  function call overhead