Create generic object pool and use it in cxx parser
Performance problems in most of the parsers are caused by calling
malloc() and free() during parsing (usually for tokens or strings)
which are expensive. Instead of destroying and creating objects
of the same time all the time, unused objects can be returned
to a pool and newly needed objects can be picked from the pool
which can avoid the allocation calls completely.

The cxx parser already uses token pool. However, its performance
suffers from one simple oversight - unused objects are returned
to the end of the list while newly requested objects are requested
from the beginning of the list. Since the pool is huge, the
elements at the beginning of the list are old and not any more
in the CPU cache and have to be fetched into the cache which
is expensive.

A simple 1-line patch would fix the above problem but as we have
the generic pool now, it's cleaner to use it instead. Being
LIFO, it doesn't suffer from the above performance problem and
increases speed from 25MB/s to 33MB/s for linux kernel C sources.